+++ src/nomadic/summarize/compute.py
@@ -15,9 +15,16 @@ variants_group_columns = [
 
 
 def filter_false_positives(variants_df: pd.DataFrame):
-    mut = variants_df[variants_df["gt_int"].isin([1,2])]
-    df = variants_df.merge(mut.groupby(variants_group_columns).agg(n_mut=pd.NamedAgg("gt_int", len), wsaf_max=pd.NamedAgg("wsaf", "max")), on=variants_group_columns)
-    df = df[~(df["n_mut"].le(1) & df["wsaf_max"].lt(0.15))].drop(columns=["n_mut", "wsaf_max"])
+    mut = variants_df[variants_df["gt_int"].isin([1, 2])]
+    df = variants_df.merge(
+        mut.groupby(variants_group_columns).agg(
+            n_mut=pd.NamedAgg("gt_int", len), wsaf_max=pd.NamedAgg("wsaf", "max")
+        ),
+        on=variants_group_columns,
+    )
+    df = df[~(df["n_mut"].le(1) & df["wsaf_max"].lt(0.15))].drop(
+        columns=["n_mut", "wsaf_max"]
+    )
     return df
